requirements.txt
langchain
langchain-openai
langchain_community
langchain_core
langgraph
langsmith
langchain-google-community
IPython
python-dotenv